Movie search timer moved to VM from List Fragment
